+++ convex/lib/publicRouteReservations.ts
@@ -1,6 +1,14 @@
 import { OPENCLAW_EXTENSION_SLUG_TO_PACKAGE } from "clawhub-schema";
 
-const RESERVED_PUBLIC_OWNER_HANDLES = new Set(["admin", "clawhub", "docs", "plugins", "skills"]);
+const RESERVED_PUBLIC_OWNER_HANDLES = new Set([
+  "admin",
+  "clawhub",
+  "creators",
+  "docs",
+  "plugins",
+  "publishers",
+  "skills",
+]);
 const RESERVED_OPENCLAW_EXTENSION_HANDLES = new Set(
   Object.keys(OPENCLAW_EXTENSION_SLUG_TO_PACKAGE),
 );
