+++ src/routes/users/index.tsx
@@ -2,6 +2,6 @@ import { createFileRoute, redirect } from "@tanstack/react-router";
 
 export const Route = createFileRoute("/users/")({
   beforeLoad: ({ search }) => {
-    throw redirect({ to: "/publishers", search, replace: true });
+    throw redirect({ to: "/creators", search, replace: true });
   },
 });
